+++ src/Mailtrap.php
@@ -2,19 +2,38 @@
 
 namespace SternerStuffWordPress;
 
+use SternerStuffWordPress\Interfaces\ActionHookSubscriber;
+
 use function Env\env;
 
 /**
  * If Mailtrap variables are set and it's on the development environment,
  * use Mailtrap for emails
  */
-class Mailtrap {
-	function __construct() {
-		add_action('phpmailer_init', [$this, 'enable_mailtrap']);
+class Mailtrap implements ActionHookSubscriber {
+
+    public static function get_actions() 
+    {
+        return [
+            'phpmailer_init' => 'phpmailer_init',
+        ];
 	}
 
-	public function enable_mailtrap($phpmailer) {
-		if(!env('MAILTRAP_USER') || !env('MAILTRAP_PASSWORD') || env('WP_ENV') != 'development') {
+    public function phpmailer_init( $phpmailer )
+    {
+        switch(env( 'MAIL_MAILER' )) {
+            case 'mailtrap':
+                $this->enable_mailtrap( $phpmailer );
+                break;
+            case 'mailhog':
+            default:
+                $this->enable_mailhog( $phpmailer );
+                break;
+        }
+    }
+
+	public function enable_mailtrap( $phpmailer ) {
+		if(!env('MAILTRAP_USER') || !env('MAILTRAP_PASSWORD')) {
 			return;
 		}
 		$phpmailer->isSMTP();
@@ -23,5 +42,14 @@ class Mailtrap {
 		$phpmailer->Port = 2525;
 		$phpmailer->Username = env('MAILTRAP_USER');
 		$phpmailer->Password = env('MAILTRAP_PASSWORD');
-	}
+    }
+    
+    public function enable_mailhog( $phpmailer )
+    {
+        $phpmailer->isSMTP();
+        $phpmailer->Host = 'localhost';
+        $phpmailer->Port = 1025;
+        $phpmailer->Username = null;
+        $phpmailer->Password = null;
+    }
 }